Screenshot/requirements.txt
ffmpeg-python==0.2.0
numba==0.54.1
numpy==1.20.3
opencv-python==4.5.4.60
Pillow==8.4.0
PyAudio
pynput==1.7.5
PyQt5==5.15.4
